+++ pages/pg_children.py
@@ -1,5 +1,5 @@
 from libs import CurrentExecution, playwright_ops
-from libs.constants import element_actions, element_properties, wait_time
+from libs.generic_constants import element_actions, element_properties, wait_time
 from libs.wrappers import *
 from pages import pg_dashboard
