+++ pages/pg_programmes.py
@@ -1,7 +1,6 @@
-from math import trunc
-
 from libs import CurrentExecution, file_ops, playwright_ops, testdata_ops
-from libs.constants import element_actions, element_properties, file_limit, wait_time
+from libs.generic_constants import element_actions, element_properties, wait_time
+from libs.mavis_constants import record_limit
 from libs.wrappers import *
 from pages import pg_sessions
 
@@ -112,7 +111,7 @@ class pg_programmes:
         self.click_continue()
         self.record_upload_time()
         wait(timeout=wait_time.MED)
-        if self.ce.get_file_record_count() > file_limit.FILE_RECORD_LIMIT:
+        if self.ce.get_file_record_count() > record_limit.FILE_RECORD_MAX_THRESHOLD:
             self.click_uploaded_file_datetime()
         self.verify_upload_output(file_path=_output_file_path)
 
@@ -122,7 +121,7 @@ class pg_programmes:
         self.click_continue()
         self.record_upload_time()
         wait(timeout=wait_time.MED)
-        if self.ce.get_file_record_count() > file_limit.FILE_RECORD_LIMIT:
+        if self.ce.get_file_record_count() > record_limit.FILE_RECORD_MAX_THRESHOLD:
             self.click_uploaded_file_datetime(truncated=True)
         self.verify_upload_output(file_path=_output_file_path)
 
@@ -135,7 +134,7 @@ class pg_programmes:
         self.click_continue()
         self.record_upload_time()
         wait(timeout=wait_time.MED)
-        if self.ce.get_file_record_count() > file_limit.FILE_RECORD_LIMIT:
+        if self.ce.get_file_record_count() > record_limit.FILE_RECORD_MAX_THRESHOLD:
             self.click_uploaded_file_datetime()
         self.verify_upload_output(file_path=_output_file_path)
