+++ pages/pg_school_moves.py
@@ -1,5 +1,9 @@
 from libs import CurrentExecution, playwright_ops
-from libs.constants import element_actions, element_properties, escape_characters
+from libs.generic_constants import (
+    element_actions,
+    element_properties,
+    escape_characters,
+)
 from libs.wrappers import *
