+++ test_data/class_list/i_year_group.csv
@@ -1,0 +1,3 @@
+CHILD_FIRST_NAME,CHILD_LAST_NAME,CHILD_PREFERRED_FIRST_NAME,CHILD_PREFERRED_LAST_NAME,CHILD_DATE_OF_BIRTH,CHILD_NHS_NUMBER,CHILD_GENDER,CHILD_ADDRESS_LINE_1,CHILD_ADDRESS_LINE_2,CHILD_TOWN,CHILD_REGISTRATION,CHILD_POSTCODE,PARENT_1_NAME,PARENT_1_RELATIONSHIP,PARENT_1_EMAIL,PARENT_1_PHONE,PARENT_2_NAME,PARENT_2_RELATIONSHIP,PARENT_2_EMAIL,PARENT_2_PHONE,CHILD_YEAR_GROUP
+C<<FNAME>>,C<<LNAME>>,PrefFirst1,PrefLast1,20101101,<<NHS_NO>>,Male,Addr1,Addr2,Town,8T5,AA1 1AA,Parent1,Dad,dad@example.com,,Parent2,Mum,mum@example.com,,8
+C<<FNAME>>,C<<LNAME>>,PrefFirst1,PrefLast1,20101101,<<NHS_NO>>,Male,Addr1,Addr2,Town,8T5,AA1 1AA,Parent1,Dad,dad@example.com,,Parent2,Mum,mum@example.com,,9
