+++ tests/test_08_vaccs_upload.py
@@ -1,6 +1,6 @@
 import pytest
 
-from libs.constants import test_data_file_paths
+from libs.mavis_constants import test_data_file_paths
 from pages import pg_dashboard, pg_import_records, pg_login, pg_programmes, pg_sessions
