+++ tests/test_50_e2e.py
@@ -1,7 +1,7 @@
 import pytest
 
 from libs import CurrentExecution
-from libs.constants import test_data_file_paths
+from libs.mavis_constants import test_data_file_paths
 from pages import pg_dashboard, pg_login, pg_programmes, pg_sessions
